Remove styled components completely
