+++ package.json
@@ -1,13 +1,16 @@
 {
   "name": "ml-classifier-ui",
+  "repository": "https://github.com/thekevinscott/ml-classifier-ui",
   "version": "0.1.3",
   "description": "A package for creating image-based machine learning models",
   "main": "index.js",
   "typings": "index.d.ts",
   "scripts": {
-    "clean": "rimraf index.js && rimraf index.d.ts && rimraf index.css && rimraf utils && rimraf Preview && rimraf Model && rimraf Dropzone",
+    "clean": "npm run clean:generated_files && npm run lean:style_types",
+    "clean:generated_files": "rimraf index.js index.d.ts index.css utils Preview Model Dropzone",
+    "clean:style_types": "rimraf src/**/*.css.d.ts src/**/*.scss.d.ts",
     "watch": "npm run build -- --watch",
-    "build": "rollup -c",
+    "build": "npm run clean:style_types && rollup -c",
     "example": "node example/server.js",
     "test": "jest"
   },
@@ -73,7 +76,6 @@
   },
   "dependencies": {
     "ml-classifier": "^0.3.1",
-    "react-spinners": "^0.3.2",
-    "styled-components": "^3.3.3"
+    "react-spinners": "^0.3.2"
   }
 }
