+++ src/App/styles.scss
@@ -1,0 +1,23 @@
+.classifier {
+  max-width: 300px;
+  background: white;
+  height: 300px;
+
+  h1,h2,h3,h4,h5,h6 {
+    font-weight: inherit;
+  }
+
+  button {
+    cursor: pointer;
+  }
+
+  ul, li {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+  }
+
+  * {
+    box-sizing: border-box;
+  }
+}
