+++ src/Model/Evaluator/Predictions/index.tsx
@@ -1,22 +1,8 @@
 import * as React from 'react';
-import styled, { StyledFunction } from "styled-components"
+import styles from './styles.scss';
 import Prediction, {
   IPrediction,
 } from './Prediction';
-
-interface ContainerProps {
-  children?: any;
-}
-
-const ul: StyledFunction<ContainerProps & React.HTMLProps<HTMLInputElement>> = styled.ul;
-
-const PredictionsContainer = ul `
-  img {
-    width: 100%;
-    max-width: 300px;
-    display: block;
-  }
-`;
 
 interface IProps {
   predictions: IPrediction[];
@@ -26,11 +12,11 @@ const Predictions: React.SFC<IProps> = ({
   predictions,
 }) => {
   return (
-    <PredictionsContainer>
+    <ul className={styles.container}>
       {predictions.map((prediction: IPrediction, idx: number) => (
         <Prediction prediction={prediction} key={idx} />
       ))}
-    </PredictionsContainer>
+    </ul>
   );
 };
