+++ src/Model/Metrics/styles.scss
@@ -1,0 +1,54 @@
+.container {
+  border-radius: 5px 5px 0 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0,13,51,0.6);
+  padding: 20px;
+  color: rgba(255,255,255,0.8);
+  font-size: 20px;
+  display: flex;
+  flex-direction: column;
+
+  button {
+    width: 120px;
+    height: 40px;
+    border: none;
+    border-radius: 5px;
+  }
+}
+
+.footer {
+  flex: 1;
+  display: flex;
+  align-items: flex-end;
+}
+
+.logs {
+  flex: 1;
+}
+
+.datum {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  flex: 1;
+
+  data {
+    font-size: 40px;
+  }
+
+  label {
+    color: rgba(255,255,255,0.6);
+  }
+}
+
+.data {
+  display: flex;
+  margin-bottom: 20px !important;
+}
+
+.header {
+  font-size: 16px;
+  margin: 0;
+}
