+++ src/Model/styles.scss
@@ -1,0 +1,4 @@
+.container {
+  display: flex;
+  flex-direction: column;
+}
