+++ src/Preview/styles.scss
@@ -1,0 +1,31 @@
+.loader {
+  opacity: 0.2;
+  position: absolute;
+  height: 100%;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
+
+.container {
+  border-radius: 5px;
+  display: flex;
+  justify-content: center;
+  position: relative;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  border: 2px dashed rgba(0,0,0,0.2);
+
+  img, canvas {
+    width: 100%;
+    height: 100%;
+    position: relative;
+  }
+
+  &.images {
+    border: 2px dashed rgba(0,0,0,0);
+  }
+}
